requirements.txt
psycopg2-binary==2.9.9
geoalchemy2==0.14.2
sqlalchemy==2.0.27
geopandas==0.14.1
shapely==2.0.2 